pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "emle-mace"
version = "0.1.0"
description = "EnergyEMLEMACE model and training CLI for EMLE QM/MM applications"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12",
    "e3nn==0.4.4",
    "numpy",
    "mace-torch",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-benchmark",
    "black",
    "isort",
]

[project.scripts]
emle-mace-train = "emle_mace.cli.run_train:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["emle_mace*"]